refactor: rename classes and remove LastDarkness
