+++ BetterUI.cs
@@ -1,6 +1,7 @@
 using BepInEx;
 using BepInEx.Configuration;
 
+using UtillI;
 using HarmonyLib;
 
 namespace BetterUI;
@@ -11,13 +12,12 @@ public class BetterUI : BaseUnityPlugin
     private static readonly Harmony harmony = new Harmony(PluginInfo.PLUGIN_GUID);
 
     public static ConfigEntry<bool> configPowerupDisplay;
-    public static ConfigEntry<bool> configInfoDisplay;
+    public static ConfigEntry<DisplayRule> configStatsPanel;
+    public static ConfigEntry<StatSelection> configStatsSelection;
 
     // config entries for info display colors
-    public static ConfigEntry<string> configheaderColor;
+    public static ConfigEntry<string> configHeaderColor;
     public static ConfigEntry<string> configLabelColor;
-
-    public static ConfigEntry<int> configLastDarkness;
 
     private void Awake()
     {
@@ -27,36 +27,33 @@ public class BetterUI : BaseUnityPlugin
             true,
             "Whether or not to show selected powerups when leveling up.");
 
-        configInfoDisplay = Config.Bind(
-            "Display.Info",
-            "DisplayInfo",
-            true,
-            "Whether or not to show current stats in pause menu.");
+        configStatsPanel = Config.Bind(
+            "Display.Stats",
+            "StatsPanelDisplayRule",
+            DisplayRule.PauseOnly,
+            "If and when to show the current stats");
 
-        configheaderColor = Config.Bind(
-            "Header.Color",
+        configStatsSelection = Config.Bind(
+            "Display.Stats",
+            "StatsSelection",
+            StatSelection.Default,
+            "Which stats to show in the panel");
+
+        configHeaderColor = Config.Bind(
+            "Display.Stats.Color",
             "HeaderColor",
             "orange",
             "What color to make the header for the stat display.");
 
         configLabelColor = Config.Bind(
-            "Label.Color",
+            "Display.Stats.Color",
             "LabelColor",
             "white",
             "What color to make the labels for the stat display.");
 
-        configLastDarkness = Config.Bind(
-            "Last.Darkness",
-            "LastDarkness",
-            1,
-            "What the darkness level should be on return to character select. (-1 to disable)");
-
-        // Plugin startup logic
+        UtillIRegister.Register(new StatsPanel());
+        harmony.PatchAll(typeof(PowerupDisplayPatch));
+        harmony.PatchAll(typeof(CloseOnEscapePatch));
         Logger.LogInfo("Better UI loaded.");
-
-        UtillI.UtillIRegister.Register(new StatsPanel());
-        harmony.PatchAll(typeof(PowerupDisplay));
-        harmony.PatchAll(typeof(LastDarkness));
-        harmony.PatchAll(typeof(CustomInputs));
     }
 }
